+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The project uses semantic versioning (see [semver](https://semver.org)).
 ## [Unreleased]
 
 
+## v0.13.2 - 2023-06-03
+
+
 ### Added
 - New `file_ownership` parameter for `stamper`, defaulting to `0664`.